affiliate_listeners_clean/requirements.txt
# Core Web3 and blockchain dependencies
web3>=6.0.0
eth-abi>=4.0.0
eth-account>=0.8.0
eth-utils>=2.0.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.24.0

# HTTP requests and async
requests>=2.28.0
aiohttp>=3.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.64.0
